ai-transcription/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.5
python-socketio>=5.10.0
speechrecognition>=3.10.0
pydub>=0.25.1
numpy>=1.26.0
websockets>=12.0
deep-translator>=1.11.4
python-dotenv>=1.0.0
